refactor(Screen): move home and detail screen to UI package
