+++ app/src/main/java/com/augieafr/todo_app/ui/component/Todo.kt
@@ -29,7 +29,7 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextDecoration
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.augieafr.todo_app.home.HomeScreenEvent
+import com.augieafr.todo_app.ui.home.HomeScreenEvent
 import com.augieafr.todo_app.ui.model.ToDoDeadline
