+++ app/src/main/java/com/augieafr/todo_app/ui/home/HomeViewModel.kt
@@ -1,4 +1,4 @@
-package com.augieafr.todo_app.home
+package com.augieafr.todo_app.ui.home
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
